% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/based_on_coordinates.R
\name{create_bbox}
\alias{create_bbox}
\title{Create a bbox}
\usage{
create_bbox(lat, long)
}
\arguments{
\item{lat}{A vector of length 2 of the latitudes}

\item{long}{A vector of length 2 of the longitudes}
}
\value{
A 2x2 matrix that can be used for `make_sexy_map`
}
\description{
Crop the bbox as retrieved by `osmdata::getbb` so
  you can zoom in.
}
\examples{
create_bbox(c(4.438865, 4.524072),
           c(52.118952, 52.184626))
}
