% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/based_on_coordinates.R
\name{crop_bbox}
\alias{crop_bbox}
\title{Crop the bbox}
\usage{
crop_bbox(bbox, left = 0, right = 0, top = 0, bottom = 0)
}
\arguments{
\item{bbox}{A 2x2 matrix, returned by `osmdata::getbb`.}

\item{left, right, top, bottom}{The proportion to crop from
with respect to this side of the bbox. Cannot be
smaller than 0 or exceed 1.}
}
\value{
A 2x2 matrix that is cropped.
}
\description{
Crop the bbox as retrieved by `osmdata::getbb` so
  you can zoom in.
}
\examples{
make_sexy_map("Leiden", "Leiden", monochrome = TRUE)
# not so nice, we just want the old city center
# not all the outskirts
leiden_bbox <- osmdata::getbb("Leiden")
leiden_cropped <- crop_bbox(leiden_bbox, .45, .15, .25, .5)
make_sexy_map(leiden_cropped, "Leiden", monochrome = TRUE)
}
